fix: handle syntax errors in config file gracefully
- Add `_load_config` helper in `config.py` to catch TOML parsing errors.
- Default to empty configuration and log an error instead of crashing.
- Ensure `load_cache_config` falls back to default memory cache if configuration is invalid or missing.
- Add unit tests for configuration robustness in `tests/unit/config/test_robustness.py`.
- Clean up build artifacts, generated version file, and bytecode.
